+++ src/pages/auth/SignInPage.tsx
@@ -1,0 +1,11 @@
+import { Screen } from "@/apps/Screen";
+
+import { NavPrevious } from "@/shared/components";
+
+export default function SignInPage() {
+    return (
+        <Screen>
+            <NavPrevious />
+        </Screen>
+    );
+}
